resolve naming error for Number_guessing_Game
